# nebus-tz/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/organizations/building/{buildingId}": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Get organizations by building ID",
                "description": "Get organizations that belong to a building based on its ID",
                "operationId": "getOrganizationsByBuilding",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "Building ID",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organizations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Building not found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/organizations/activity/{activityId}": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Get organizations by activity ID",
                "description": "Get organizations that belong to an activity based on its ID",
                "operationId": "getOrganizationsByActivity",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "Activity ID",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organizations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Activity not found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/organization/{id}": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Get organization by ID",
                "description": "Get the details of an organization by its ID",
                "operationId": "getOrganization",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "Organization ID",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organization retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organization"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Organization not found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/organizations/near": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Get organizations in a specified radius from a point",
                "description": "Get a list of organizations within a specified radius (in kilometers) from a given latitude and longitude",
                "operationId": "getOrganizationsInRadius",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "description": "Latitude",
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "description": "Longitude",
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "description": "Radius in kilometers",
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organizations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/organizations/search/{name}": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Search organizations by name",
                "description": "Search organizations by a name (partial matching)",
                "operationId": "searchOrganizationsByName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "Organization name",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organizations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No organizations found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/organizations/activity-name": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Get organizations by activity name",
                "description": "Get organizations that belong to an activity based on its name",
                "operationId": "getOrganizationsByActivityName",
                "parameters": [
                    {
                        "name": "activity_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "description": "Activity name",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organizations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Activity not found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Organization": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Organization ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Organization Name",
                        "type": "string"
                    },
                    "building_id": {
                        "description": "Building ID",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Creation date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "API Key for authentication",
                "name": "Api-key-custom",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Organization",
            "description": "Organization"
        }
    ]
}